@@ -2,11 +2,15 @@
   "name": "0x40-web",
   "version": "5.4",
   "description": "Pretty images and colours",
-  "main": "index.html",
+  "type": "module",
   "directories": {
     "doc": "docs"
   },
   "scripts": {
+    "dev": "vite",
+    "build": "vite build",
+    "preview": "vite preview",
+    "check": "svelte-check --tsconfig ./tsconfig.app.json && tsc -p tsconfig.node.json",
     "format": "prettier --write ./**/*.svelte ./**/*.css ./**/*.ts"
   },
   "repository": {
@@ -20,29 +24,21 @@
   },
   "homepage": "https://github.com/mon/0x40-web#readme",
   "devDependencies": {
-    "@tsconfig/svelte": "^3.0.0",
-    "css-loader": "^6.7.3",
-    "esbuild-loader": "^3.0.1",
-    "mini-css-extract-plugin": "^2.7.3",
-    "prettier": "^2.8.4",
-    "prettier-plugin-svelte": "^2.9.0",
-    "source-map-loader": "^4.0.1",
-    "style-loader": "^3.3.1",
-    "svelte": "^3.56.0",
-    "svelte-check": "^3.1.2",
-    "svelte-check-plugin": "^1.0.4",
-    "svelte-loader": "^3.1.5",
-    "svelte-preprocess": "^5.0.1",
-    "ts-loader": "^9.4.2",
-    "typescript": "^4.9.5",
-    "webpack": "^5.76.1",
-    "webpack-cli": "^5.0.1",
-    "webpack-dev-server": "^4.11.1"
+    "@sveltejs/vite-plugin-svelte": "^5.0.3",
+    "@tsconfig/svelte": "^5.0.4",
+    "@types/serve-static": "^1.15.7",
+    "prettier": "^3.5.3",
+    "prettier-plugin-svelte": "^3.3.3",
+    "serve-static": "^2.2.0",
+    "svelte": "^5.28.2",
+    "svelte-check": "^4.1.6",
+    "typescript": "^5.8.3",
+    "vite": "^6.3.4"
   },
   "dependencies": {
     "@types/string_score": "^0.1.28",
     "@wasm-audio-decoders/ogg-vorbis": "^0.1.16",
-    "@zip.js/zip.js": "^2.6.78",
+    "@zip.js/zip.js": "^2.7.60",
     "codec-parser": "^2.5.0",
     "mpg123-decoder": "^1.0.0",
     "ogg-opus-decoder": "^1.6.14",
