@@ -67,86 +67,94 @@
 
     <div>Songs: {pack.songs.length}</div>
     <table>
-      <tr>
-        <th>Title</th>
-        <th>Source</th>
-        <th>Banks</th>
-        <th>Loop len</th>
-        <th>Loop file</th>
-        <th>Build len</th>
-        <th>Build file</th>
-        <th>charsPerBeat</th>
-        <th>independentBuild</th>
-      </tr>
-      {#each pack.songs as song}
+      <thead>
         <tr>
-          <td><input type="text" bind:value={song.title} /></td>
-          <td><input type="text" bind:value={song.source} /></td>
-          <td>{song.bankCount}</td>
-          <td>{song.loop.mapLen}</td>
-          <td>{song.loop.filename}</td>
-          <td>{song.build ? song.build.mapLen : "n/a"}</td>
-          <td>{song.build ? song.build.filename : "n/a"}</td>
-          <td>{song.charsPerBeat}</td>
-          <td>
-            {#if song.build}
-              <input type="checkbox" bind:value={song.independentBuild} />
-            {:else}
-              n/a
-            {/if}
-          </td>
+          <th>Title</th>
+          <th>Source</th>
+          <th>Banks</th>
+          <th>Loop len</th>
+          <th>Loop file</th>
+          <th>Build len</th>
+          <th>Build file</th>
+          <th>charsPerBeat</th>
+          <th>independentBuild</th>
         </tr>
-      {/each}
+      </thead>
+      <tbody>
+        {#each pack.songs as song}
+          <tr>
+            <td><input type="text" bind:value={song.title} /></td>
+            <td><input type="text" bind:value={song.source} /></td>
+            <td>{song.bankCount}</td>
+            <td>{song.loop.mapLen}</td>
+            <td>{song.loop.filename}</td>
+            <td>{song.build ? song.build.mapLen : "n/a"}</td>
+            <td>{song.build ? song.build.filename : "n/a"}</td>
+            <td>{song.charsPerBeat}</td>
+            <td>
+              {#if song.build}
+                <input type="checkbox" bind:value={song.independentBuild} />
+              {:else}
+                n/a
+              {/if}
+            </td>
+          </tr>
+        {/each}
+      </tbody>
     </table>
 
     <div>Images: {pack.images.length}</div>
     <table>
-      <tr>
-        <th>Name</th>
-        <th>Full name</th>
-        <th>Source</th>
-        <th>Align</th>
-        <th>Img #</th>
-        <th>frameDurations</th>
-        <th>beatsPerAnim</th>
-        <th>syncOffset</th>
-      </tr>
-      {#each pack.images as image}
+      <thead>
         <tr>
-          <td><input type="text" bind:value={image.name} /></td>
-          <td><input type="text" bind:value={image.fullname} /></td>
-          <td><input type="text" bind:value={image.source} /></td>
-          <td>
-            <select bind:value={image.align}>
-              <option>center</option>
-              <option>left</option>
-              <option>right</option>
-            </select>
-          </td>
-          <td>{image.bitmaps.length}</td>
-          <td>
-            {#if image.animated}
-              <input type="text" bind:value={image.rawFrameDurations} />
-            {:else}
-              n/a
-            {/if}
-          </td>
-          <td>
-            {#if image.animated}
-              <input type="number" bind:value={image.beatsPerAnim} />
-            {:else}
-              n/a
-            {/if}
-          </td>
-          <td>
-            {#if image.animated}
-              <input type="number" bind:value={image.syncOffset} />
-            {:else}
-              n/a
-            {/if}
-          </td>
+          <th>Name</th>
+          <th>Full name</th>
+          <th>Source</th>
+          <th>Align</th>
+          <th>Img #</th>
+          <th>frameDurations</th>
+          <th>beatsPerAnim</th>
+          <th>syncOffset</th>
         </tr>
-      {/each}
+      </thead>
+      <tbody>
+        {#each pack.images as image}
+          <tr>
+            <td><input type="text" bind:value={image.name} /></td>
+            <td><input type="text" bind:value={image.fullname} /></td>
+            <td><input type="text" bind:value={image.source} /></td>
+            <td>
+              <select bind:value={image.align}>
+                <option>center</option>
+                <option>left</option>
+                <option>right</option>
+              </select>
+            </td>
+            <td>{image.bitmaps.length}</td>
+            <td>
+              {#if image.animated}
+                <input type="text" bind:value={image.rawFrameDurations} />
+              {:else}
+                n/a
+              {/if}
+            </td>
+            <td>
+              {#if image.animated}
+                <input type="number" bind:value={image.beatsPerAnim} />
+              {:else}
+                n/a
+              {/if}
+            </td>
+            <td>
+              {#if image.animated}
+                <input type="number" bind:value={image.syncOffset} />
+              {:else}
+                n/a
+              {/if}
+            </td>
+          </tr>
+        {/each}
+      </tbody>
     </table>
   </div>
 {:else}
